chore: selenium-jupiter update
