+++ vaadin-testbench-core-junit5/src/main/java/com/vaadin/testbench/BrowserTestBase.java
@@ -10,6 +10,7 @@
 package com.vaadin.testbench;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.HasCapabilities;
 import org.openqa.selenium.WebDriver;
@@ -18,32 +19,20 @@ import org.openqa.selenium.WebDriver;
  * A superclass with helper methods to aid TestBench developers create a JUnit 5
  * based tests.
  */
-public abstract class BrowserTestBase extends AbstractBrowserTestBase
+public abstract class BrowserTestBase extends AbstractBrowserDriverTestBase
         implements HasCapabilities {
 
-    private WebDriver driver;
+    @RegisterExtension
+    public ScreenshotOnFailureExtension screenshotOnFailureExtension = new ScreenshotOnFailureExtension(
+            this, true);
 
     private Capabilities capabilities;
 
     @BeforeEach
     public void setWebDriverAndCapabilities(WebDriver driver,
             Capabilities capabilities) {
-        this.driver = driver;
+        setDriver(driver);
         this.capabilities = capabilities;
-    }
-
-    /**
-     * Returns the {@link WebDriver} instance previously specified within
-     * {@link #setWebDriverAndCapabilities(WebDriver, Capabilities)}, or (if the
-     * previously provided WebDriver instance was not already a
-     * {@link TestBenchDriverProxy} instance) a {@link TestBenchDriverProxy}
-     * that wraps that driver.
-     *
-     * @return the active WebDriver instance
-     */
-    @Override
-    public WebDriver getDriver() {
-        return driver;
     }
 
     @Override
