+++ vaadin-testbench-integration-tests-junit5/src/test/java/com/vaadin/tests/SeleniumPageObjectIT.java
@@ -2,32 +2,17 @@ package com.vaadin.tests;
 
 import java.util.List;
 
-import io.github.bonigarcia.seljup.DriverCapabilities;
-import io.github.bonigarcia.seljup.SeleniumJupiter;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.testUI.PageObjectView;
 
-@ExtendWith(SeleniumJupiter.class)
-public class SeleniumPageObjectIT extends SeleniumAbstractTB9Test {
-
-    @DriverCapabilities
-    ChromeOptions options = new ChromeOptions();
+public class SeleniumPageObjectIT extends AbstractSeleniumChromeTB9Test {
 
     @Override
     protected Class<? extends Component> getTestView() {
         return PageObjectView.class;
-    }
-
-    @BeforeEach
-    public void setDriver(ChromeDriver driver) {
-        super.setDriver(driver);
     }
 
     @Test
